+++ src/receiverFMQ.cxx
@@ -187,8 +187,8 @@ int main(int argc, const char **argv) {
           // first part is STF header
 
           int i = 0;
-	  bool dumpNext = false;
-	  SubTimeframe *stf = nullptr;
+          bool dumpNext = false;
+          SubTimeframe *stf = nullptr;
           for (auto const &mm : msgParts) {
             if (i == 0) {
               if (mm->GetSize() != sizeof(SubTimeframe)) {
@@ -198,7 +198,7 @@ int main(int argc, const char **argv) {
                 break;
               }
               stf = (SubTimeframe *)mm->GetData();
-              if (stf->numberOfHBF != nPart - 1) {
+              if ((int)stf->numberOfHBF != nPart - 1) {
                 theLog.log(
                     InfoLogger::Severity::Error,
                     "Mismatch stf->numberOfHBF %d != %d message parts - 1\n",
@@ -208,8 +208,8 @@ int main(int argc, const char **argv) {
               // number of message parts matches number of HBFs in header ?
               if (cfgDumpTF) {
                 if ((stf->timeframeId == 1) ||
-                    (stf->timeframeId % cfgDumpTF == 0)) {                  
-	          dumpNext = true;
+                    (stf->timeframeId % cfgDumpTF == 0)) {
+                  dumpNext = true;
                 }
               }
             } else {
@@ -226,12 +226,12 @@ int main(int argc, const char **argv) {
 
                 if (dumpNext) {
                   printf("Receiving TF %d CRU %d link %d : %d HBf\n",
-                         (int)stf->timeframeId, (int)h.getCruId(), (int)stf->linkId,
-                         (int)stf->numberOfHBF);
-		  dumpNext = false;
-		}
-		
-		if (cfgDumpRDH) {
+                         (int)stf->timeframeId, (int)h.getCruId(),
+                         (int)stf->linkId, (int)stf->numberOfHBF);
+                  dumpNext = false;
+                }
+
+                if (cfgDumpRDH) {
                   h.dumpRdh(pageOffset, 1);
                 }
 
@@ -290,8 +290,10 @@ int main(int argc, const char **argv) {
   theLog.log("Receiving loop completed");
   theLog.log(
       "bytes received: %llu  (avg=%.2lf  min=%llu  max=%llu  count=%llu)",
-      msgStats.get(), msgStats.getAverage(), msgStats.getMinimum(),
-      msgStats.getMaximum(), msgStats.getCount());
+      (unsigned long long)msgStats.get(), msgStats.getAverage(),
+      (unsigned long long)msgStats.getMinimum(),
+      (unsigned long long)msgStats.getMaximum(),
+      (unsigned long long)msgStats.getCount());
 
   return 0;
